Crianda a interacao entre back e front

+++ backend/backend/src/main/java/com/bmss/backend/repository/CommentRepository.java
@@ -3,5 +3,4 @@ package com.bmss.backend.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.bmss.backend.model.Comment;
 
-public interface CommentRepository extends JpaRepository<Comment, Long> {
-}
+public interface CommentRepository extends JpaRepository<Comment, Integer> {}
